Started using express router, added api for adding new user,getting user,updating users array of books
