+++ config/config.env
@@ -1,2 +1,2 @@
 PORT=5000
-MONGO_URI=mongodb+srv://tarik-besic:{pass}@cluster0.pk5um.mongodb.net/user-data?retryWrites=true&w=majority
+MONGO_URI=mongodb+srv://tarik-besic:Parkourislife2424@cluster0.pk5um.mongodb.net/user-data?retryWrites=true&w=majority
