+++ config/database.js
@@ -6,7 +6,8 @@ const connectDB= async()=>{
     try{
         const conn=await mongoose.connect(process.env.MONGO_URI,{
          useNewUrlParser: true,
-         useUnifiedTopology: true
+         useUnifiedTopology: true,
+         useFindAndModify:false
         });
         console.log(`Database connected ${conn.connection.host}`);
     }
