+++ views/createClass.ejs
@@ -9,7 +9,7 @@
     <%- include('./partials/nav.ejs') %>
     
 
-<form method="POST" action="/upload" enctype="multipart/form-data">
+<form method="POST" action="/class/upload" enctype="multipart/form-data">
     <input type="file" name="upload" accept=".xls,.xlsx,.ods">
     <a href="#"><button class="button btn-primary">SUBMIT</button></a>
 </form>
