+++ views/users.ejs
@@ -13,7 +13,7 @@
     
     <%- include('./partials/nav.ejs') %>
     
-    <form action="/getusers/query/" method="get">
+    <form action="users/getusers/" method="get">
     <div class="container mt-5">
         <div class="row d-flex justify-content-center">
             <div class="col-md-10">
